Move trace list pack method to `pack` module
Wrap up all packing facilities as part of the `pack` module.
