+++ src/trace.rs
@@ -1,13 +1,11 @@
-use std::ffi::{
-    c_char, c_double, c_float, c_int, c_long, c_uchar, c_uint, c_ulong, c_void, CString,
-};
+use std::ffi::{c_char, c_double, c_float, c_int, c_long, c_uchar, c_uint, c_ulong};
 use std::fmt;
 use std::ptr;
 use std::slice::from_raw_parts;
 
 use crate::{
-    error::check, pack, raw, util, MSControlFlags, MSDataEncoding, MSError, MSRecord, MSResult,
-    MSSampleType, MSSubSeconds, MSTimeFormat,
+    error::check, raw, util, MSControlFlags, MSError, MSRecord, MSResult, MSSampleType,
+    MSSubSeconds, MSTimeFormat,
 };
 use time::OffsetDateTime;
 
@@ -448,61 +446,6 @@ impl MSTraceList {
         Ok(())
     }
 
-    /// Packs the trace lists' data into miniSEED records.
-    ///
-    /// Buffers containing the packed miniSEED records are passed to the `record_handler` closure.
-    /// Returns on success a tuple where the first value is the number of totally packed records
-    /// and the second value is the number of totally packed samples.
-    ///
-    /// Packing is controlled by the following `flags`:
-    /// - If `flags` has [`MSControlFlags::MSF_FLUSHDATA`] set, all of the trace lists' data will be
-    /// packed into miniSEED records even though the last one will probably be smaller than
-    /// requested or, in the case of miniSEED v2, unfilled.
-    /// - If `flags` has [`MSControlFlags::MSF_PACKVER2`] set records are packed as miniSEED v2.
-    /// - If `flags` has [`MSControlFlags::MSF_MAINTAINMSTL`] packed data is not removed from the
-    /// trace lists' internal buffers.
-    ///
-    /// See also [`pack()`] for packing raw data samples.
-    pub fn pack<F>(
-        &mut self,
-        mut record_handler: F,
-        info: &TlPackInfo,
-        flags: MSControlFlags,
-    ) -> MSResult<(c_long, c_long)>
-    where
-        F: FnMut(&[u8]),
-    {
-        let mut extra_ptr = ptr::null_mut();
-        if let Some(extra_headers) = &info.extra_headers {
-            let cloned = extra_headers.clone();
-            extra_ptr = cloned.into_raw();
-        }
-
-        let mut cnt_samples: c_long = 0;
-        let cnt_samples_ptr = &mut cnt_samples as *mut _;
-        let cnt_records = unsafe {
-            check(raw::mstl3_pack(
-                self.inner,
-                Some(pack::rh_wrapper::<F>),
-                (&mut record_handler) as *mut _ as *mut c_void,
-                info.rec_len,
-                info.encoding as c_char,
-                cnt_samples_ptr,
-                flags.bits(),
-                0,
-                extra_ptr,
-            ))?
-        };
-
-        if !extra_ptr.is_null() {
-            unsafe {
-                let _ = CString::from_raw(extra_ptr);
-            }
-        }
-
-        Ok((cnt_records, cnt_samples))
-    }
-
     /// Returns an object that implements [`Display`] for printing a trace list summary.
     ///
     /// By default only prints the FDSN source identifier, starttime and endtime for each
@@ -534,34 +477,6 @@ impl MSTraceList {
 impl Drop for MSTraceList {
     fn drop(&mut self) {
         unsafe { raw::mstl3_free((&mut self.inner) as *mut *mut MS3TraceList, 1) };
-    }
-}
-
-/// Struct aggregating [`MSTraceList`] packing information.
-///
-/// See also [`PackInfo`](crate::PackInfo).
-#[derive(Debug, Clone)]
-pub struct TlPackInfo {
-    // /// The miniSEED format version.
-    // pub format_version: c_uchar,
-    /// Data encoding.
-    pub encoding: MSDataEncoding,
-    /// Record length used for encoding.
-    pub rec_len: c_int,
-    /// Extra headers.
-    ///
-    /// If not `None` it is expected to contain extra headers, i.e. a string containing (compact)
-    /// JSON, that will be added to each output record.
-    pub extra_headers: Option<CString>,
-}
-
-impl Default for TlPackInfo {
-    fn default() -> Self {
-        Self {
-            encoding: MSDataEncoding::Steim2,
-            rec_len: 4096,
-            extra_headers: None,
-        }
     }
 }
